server/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
supabase
pydantic[email]
pydantic
sqlalchemy
psycopg[binary]

